add zlib middleware
