+++ internal/server/controller/handler.go
@@ -2,13 +2,13 @@ package controller
 
 import (
 	"github.com/gin-gonic/gin"
-	"github.com/zelas91/metric-collector/internal/server/middleware"
+	"github.com/zelas91/metric-collector/internal/server/controller/middleware"
 )
 
 func (h *MetricHandler) InitRoutes() *gin.Engine {
 	router := gin.New()
 
-	router.Use(middleware.WithLogging)
+	router.Use(middleware.WithLogging, middleware.GzipMiddleware)
 	router.GET("/", h.GetMetrics)
 	update := router.Group("/update")
 	value := router.Group("/value")
